# src/api/v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ferni Developer Platform API",
    "description": "Extend Ferni agents with custom workflows, external services, MCP servers, and webhooks.",
    "version": "2.0.0",
    "contact": {
      "name": "Ferni Developer Support",
      "url": "https://developers.ferni.ai",
      "email": "developers@ferni.ai"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://ferni.ai/terms"
    }
  },
  "servers": [
    {
      "url": "https://api.ferni.ai/api/v2",
      "description": "Production"
    },
    {
      "url": "http://localhost:3002/api/v2",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/developers/mcp-servers/{serverId}/test": {
      "post": {
        "tags": ["MCP Servers"],
        "summary": "Test MCP server connection",
        "operationId": "testMcpServer",
        "parameters": [
          { "$ref": "#/components/parameters/ServerId" }
        ],
        "responses": {
          "200": {
            "description": "Connection test result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/McpTestResult" }
              }
            }
          }
        }
      }
    },
...
    "/developers/tools/{toolId}/test": {
      "post": {
        "tags": ["Custom Tools"],
        "summary": "Test custom tool",
        "operationId": "testTool",
        "parameters": [
          { "$ref": "#/components/parameters/ToolId" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "object",
                    "description": "Test input parameters"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ToolTestResult" }
              }
            }
          }
        }
      }
    },
...
    "/developers/webhooks/{webhookId}/test": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Send test event",
        "operationId": "testWebhook",
        "parameters": [
          { "$ref": "#/components/parameters/WebhookId" }
        ],
        "responses": {
          "200": {
            "description": "Test event sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WebhookTestResult" }
              }
            }
          }
        }
      }
    },
...
    "/developers/activities": {
      "get": {
        "tags": ["Activities"],
        "summary": "List activities",
        "operationId": "listActivities",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["started", "completed", "failed"]
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of activities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "activities": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Activity" }
                    },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Activities"],
        "summary": "Log activity",
        "operationId": "createActivity",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ActivityCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Activity created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Activity" }
              }
            }
          }
        }
      }
    },
    "/developers/activities/stats": {
      "get": {
        "tags": ["Activities"],
        "summary": "Get activity stats",
        "operationId": "getActivityStats",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity statistics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityStats" }
              }
            }
          }
        }
      }
    },
    "/developers/activities/{activityId}": {
      "get": {
        "tags": ["Activities"],
        "summary": "Get activity",
        "operationId": "getActivity",
        "parameters": [
          { "$ref": "#/components/parameters/ActivityId" }
        ],
        "responses": {
          "200": {
            "description": "Activity details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Activity" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Activities"],
        "summary": "Delete activity",
        "operationId": "deleteActivity",
        "parameters": [
          { "$ref": "#/components/parameters/ActivityId" }
        ],
        "responses": {
          "204": { "description": "Activity deleted" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Publisher API key (pk_live_xxx or pk_test_xxx)"
      }
    },
    "parameters": {
      "ServerId": {
        "name": "serverId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "MCP server ID (mcp_xxx)"
      },
      "ToolId": {
        "name": "toolId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Tool ID (tool_xxx)"
      },
      "WebhookId": {
        "name": "webhookId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Webhook ID (wh_xxx)"
      },
...
      "ActivityId": {
        "name": "activityId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Activity ID (act_xxx)"
      }
    },
...
    "schemas": {
...
      "McpTestResult": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "latencyMs": { "type": "integer" },
          "toolCount": { "type": "integer" },
          "error": { "type": "string" }
        }
      },
...
      "ToolTestResult": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "result": { "type": "object" },
          "latencyMs": { "type": "integer" },
          "error": { "type": "string" }
        }
      },
...
      "WebhookTestResult": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "statusCode": { "type": "integer" },
          "latencyMs": { "type": "integer" },
          "error": { "type": "string" }
        }
      },
...
      "Activity": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "act_abc123" },
          "publisherId": { "type": "string" },
          "personaId": { "type": "string" },
          "userId": { "type": "string" },
          "sessionId": { "type": "string" },
          "type": { "type": "string" },
          "name": { "type": "string" },
          "data": { "type": "object" },
          "startedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time" },
          "duration": { "type": "integer" },
          "status": {
            "type": "string",
            "enum": ["started", "completed", "failed"]
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ActivityCreate": {
        "type": "object",
        "required": ["type", "name"],
        "properties": {
          "personaId": { "type": "string" },
          "userId": { "type": "string" },
          "sessionId": { "type": "string" },
          "type": { "type": "string" },
          "name": { "type": "string" },
          "data": { "type": "object" },
          "status": {
            "type": "string",
            "enum": ["started", "completed", "failed"],
            "default": "completed"
          }
        }
      },
      "ActivityStats": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "completed": { "type": "integer" },
          "failed": { "type": "integer" },
          "avgDuration": { "type": "number" },
          "byType": {
            "type": "object",
            "additionalProperties": { "type": "integer" }
          }
        }
      }
    }
  }
}